+++ src/components/MainPage/Main/Sections/LocalItem.tsx
