+++ src/components/MainPage/Main/Sections/LocalSection.tsx
