+++ src/components/MainPage/Main/Sections/WebFeedItem.tsx
