+++ src/components/MainPage/SearchElements/SearchFilterItem.tsx
