+++ src/components/MainPage/SearchElements/SearchIntroduction.tsx
