+++ src/components/MainPage/SearchElements/SearchItem.tsx
